requirements.txt
asgiref==3.4.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
commandlines==0.4.1
cycler==0.10.0
decorator==4.4.2
Django==3.2.4
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.11.0
filetype==1.0.7
Flask==2.0.1
hsh==1.1.0
idna==2.10
imageio==2.9.0
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
llvmlite==0.36.0
MarkupSafe==2.0.1
matplotlib==3.4.2
networkx==2.5.1
numba==0.53.1
numpy==1.21.0
Pillow==8.2.0
PyMatting==1.1.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
rembg==1.0.26
requests==2.25.1
scikit-image==0.18.1
scipy==1.7.0
six==1.16.0
sqlparse==0.4.1
tifffile==2021.6.14
torch==1.7.1+cpu
torchvision==0.8.2+cpu
tqdm==4.61.1
typing-extensions==3.10.0.0
urllib3==1.26.6
waitress==2.0.0
Werkzeug==2.0.1
